refactoring code and implemented promotionSetA discount

+++ src/main/java/com/product/flipdeal/model/CurrencyExchange.java
@@ -2,13 +2,14 @@ package com.product.flipdeal.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 public class CurrencyExchange {
     String base;
     String date;
-    Map<String, Double> rates;
+    Map<String, BigDecimal> rates;
     boolean success;
     long timestamp;
 }
